requirements.txt
carla==0.9.10
certifi==2023.5.7
charset-normalizer==3.1.0
cycler==0.11.0
decorator==5.1.1
dictor==0.1.11
einops==0.3.2
ephem==3.7.7.1
fonttools==4.40.0
geffnet==1.0.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.5.0
networkx==2.2
numpy==1.18.1
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
py-trees==0.8.3
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
scipy==1.4.1
setuptools-scm==7.1.0
shapely==2.0.1
simple-watchdog-timer==0.1.1
six==1.14.0
tabulate==0.8.7
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
